refactor: simplify GameDialogModal and Header components by extracting MatchHistoryPanel and MobileGameHeaderControls
