+++ src/lib/presentation/components/Header.svelte
@@ -12,6 +12,7 @@
   import GeneratePieceButton from "$lib/presentation/components/GeneratePieceButton.svelte";
   import Icon from "$lib/presentation/components/Icon.svelte";
   import IconButton from "$lib/presentation/components/IconButton.svelte";
+  import MobileGameHeaderControls from "$lib/presentation/components/MobileGameHeaderControls.svelte";
   import PlayerIdentityBadge from "$lib/presentation/components/PlayerIdentityBadge.svelte";
   import { DESKTOP_NAVIGATION_BREAKPOINT_PX } from "$lib/presentation/constants/UiConstants";
   import { playerDisplayName } from "$lib/presentation/matchPresentation";
@@ -240,81 +241,26 @@
       : "flex h-16 items-center px-2 py-1 lg:hidden"}
   >
     {#if isGamePage}
-      <div class="absolute top-1/2 left-1 -translate-y-1/2">
-        <IconButton icon="menu" label={m.drawer_title()} onclick={onClickMenu} />
-      </div>
-
-      <div class="absolute top-1/2 right-1 -translate-y-1/2">
-        <IconButton icon="exit_to_app" label={m.leave_match()} onclick={openLeaveDialog} />
-      </div>
-
-      <div class="flex min-h-26 flex-col items-start gap-1 px-12 py-1">
-        <div class="relative flex items-center gap-2">
-          <PlayerIdentityBadge
-            player="opponent"
-            label={opponentLabel}
-            resource={opponentResources}
-            compact={true}
-            previewCost={turn.player === Player.OPPONENT ? hoveredGenerationCost : undefined}
-            additionalClass="h-10 w-fit min-w-0 px-2 py-1"
-          />
-
-          <AppButton
-            additionalClass="mt-0 mb-0 h-10 w-10 shrink-0 rounded-2xl px-0"
-            onclick={toggleGenerateMenu}
-            variant="primary"
-          >
-            <Icon icon={generateMenuPieceTypes[0].config.iconName} size={20} />
-            <span class="sr-only">{m.produce()}</span>
-          </AppButton>
-
-          {#if isGenerateMenuOpen}
-            <div class="pointer-events-none absolute top-0 left-[calc(100%+0.5rem)] z-30">
-              <div
-                class="bg-surface/95 dark:bg-surface-dark/95 border-outline dark:border-outline-dark pointer-events-auto w-max max-w-[calc(100vw-12rem)] rounded-2xl border px-2 py-2 shadow-lg backdrop-blur-sm"
-              >
-                <div class="flex flex-wrap items-center justify-start gap-2">
-                  {#each generateMenuPieceTypes as pieceType (pieceType.config.iconName)}
-                    <GeneratePieceButton
-                      {pieceType}
-                      compact={true}
-                      onPreviewChange={handleGenerationCostPreview}
-                      additionalClass="mt-0 mb-0 h-10 w-10 rounded-2xl px-0 py-0"
-                    />
-                  {/each}
-                </div>
-              </div>
-            </div>
-          {/if}
-        </div>
-
-        <div class="grid grid-cols-[auto_auto_auto] items-center justify-start gap-2">
-          <PlayerIdentityBadge
-            player="self"
-            label={selfLabel}
-            resource={selfResources}
-            compact={true}
-            previewCost={turn.player === Player.SELF ? hoveredGenerationCost : undefined}
-            additionalClass="h-10 w-fit min-w-0 px-2 py-1"
-          />
-
-          <EndTurnButton
-            compact={true}
-            additionalClass="mt-0 mb-0 h-10 w-10 rounded-2xl px-0 py-0"
-          />
-
-          <div title={generationModeLabel} aria-label={generationModeLabel}>
-            <AppButton
-              additionalClass="mt-0 mb-0 h-10 w-10 rounded-2xl px-0 py-0"
-              onclick={toggleGenerationMode}
-              disabled={!isHumanTurn || isAutomationRunning || turn.winner !== null}
-            >
-              <Icon icon={generationModeIcon} size={20} />
-              <span class="sr-only">{generationModeLabel}</span>
-            </AppButton>
-          </div>
-        </div>
-      </div>
+      <MobileGameHeaderControls
+        {opponentLabel}
+        {opponentResources}
+        {selfLabel}
+        {selfResources}
+        currentPlayer={turn.player}
+        {hoveredGenerationCost}
+        {generationModeLabel}
+        {generationModeIcon}
+        {isGenerateMenuOpen}
+        {generateMenuPieceTypes}
+        {isHumanTurn}
+        {isAutomationRunning}
+        winner={turn.winner}
+        {onClickMenu}
+        onOpenLeaveDialog={openLeaveDialog}
+        onToggleGenerateMenu={toggleGenerateMenu}
+        onToggleGenerationMode={toggleGenerationMode}
+        onPreviewChange={handleGenerationCostPreview}
+      />
     {:else}
       <div class="flex w-full items-center justify-between">
         <IconButton icon="menu" label={m.drawer_title()} onclick={onClickMenu} />
